backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
